+++ Firestore/Source/Public/FirebaseFirestore/FIRQuery.h
@@ -579,7 +579,7 @@ NS_SWIFT_NAME(Query)
 * @return An `AggregateQuery` encapsulating this `Query` and `AggregateField`s, which can be used
  * to query the server for the aggregation results.
  */
-- (FIRAggregateQuery *)aggregate:(NSArray<FIRAggregateField *> *)aggregations NS_SWIFT_NAME(aggregate(_ aggregations:));
+- (FIRAggregateQuery *)aggregate:(NSArray<FIRAggregateField *> *)aggregations NS_SWIFT_NAME(aggregate(_:));
 
 @end