+++ Firestore/core/src/model/aggregate_field.cc
@@ -22,11 +22,6 @@ namespace firebase {
 namespace firestore {
 namespace model {
 
-bool operator==(const AggregateField& lhs, const AggregateField& rhs) {
-  return lhs.op == rhs.op && lhs.alias == rhs.alias &&
-         lhs.fieldPath.CanonicalString() == rhs.fieldPath.CanonicalString();
-}
-
 size_t AggregateField::Hash() const {
   return util::Hash(op, alias, fieldPath.CanonicalString());
 }